+++ frontend/src/hooks/useToggle.js
@@ -4,8 +4,8 @@ import React from "react";
 /**
  * A hook that returns a boolean value and a function to toggle it.
  *
- * @param {boolean} initialValue - The initial value of the boolean state.
- * @returns {[boolean, function]} - A tuple containing the boolean state and a function to toggle it.
+ * @param {boolean | (() => boolean)} initialValue - The initial value of the boolean state.
+ * @returns {[boolean, () => void]} - A tuple containing the boolean state and a function to toggle it.
  */
 function useToggle(initialValue = false) {
     if (typeof initialValue !== "boolean" && typeof initialValue !== "function") {
